Added the private tx_key into the tx_extra of the block reward tx
Added the private tx_key into the tx_extra of the block reward tx.

The reason for this is to make block reward verifiable in the X-CASH proof of stake system. This is to create a verifiable way to prove that the block validation nodes are not mining any blocks, unless it is their turn to be the block producer.