Changed how the first xcash_proof_of_stake block is verified
Changed how the first xcash_proof_of_stake block is verified, as the main network data node will create and sign only the first xcash_proof_of_stake block by itself, and this will be a valid block on the network. This is to make the transition to xcash_proof_of_stake, as their will be no VRF data in the previous block, so the main network data node will create the first xcash_proof_of_stake block with the VRF data.