Added simplfied block verification for current block verifiers
Added 2 different modes for verifying a block in XCASH_DPOPS. If the node is not a current block verifier it will sync the block from a random current block verifier and will validate the block hash and verify the block is correct and signed by at least 67% of the block verifiers.

If the node is a current block verifier, since the block verifier would have already had to verifiy that block, just confirm the data hashes.

Checks have been added for cases where a block verifier is in and then out of the top 100 and where they go offline for a few rounds. In these cases they will need to sync the missing blocks from a random block verifier before they switch to the simplfied syncing mode, since this mode will only work for live syncing.

The code has created a slight performance delay for non current block verifiers, as they will need wait 2 seconds every 288 blocks (1 days worth of blocks) to check if they are a current block verifier, so overall not much time will be added to a nodes total syncing time.