Changed how a node verifies a block
Simplified the process of verifying a block for a node and for a block verifier.

Initially the block verifiers will check the complete structure of the block, make sure the correct block producer created the block, and make sure that the valid amount of signatures are in the block.

The node was then to do the same thing minus not checking for the correct block producer. Since the whole premise of the system is based on at least 67 block verifiers agreeing, it does not make sense for every individual node to check all of the data. Instead a simplfied check of the block structure has been used as well as just comparing the hash of the block the node received and making sure at least 67 block verifiers say that is the correct hash.