feat: Added wait until valid database time for wallet functions
Added a wait time for delegate functions since votes are accepted the first 5 minutes of the hour and database changes are accepted between the 2-4 minutes of each block. Since this would be hard for a user to know, a built in wait has been added that will automatically wait until the next valid data time and then submit the wallet function.