Corrected IP_address modification
The "IP_address" modification appears to attempt to check for a decimal point "." prior to allowing the command to proceed. Instead it checks for the absence with a != before it proceeds, causing it to error if any decimal point is found. Changing the code to == npos corrects the behavior to require at least one decimal point.